The changes I have made are described in CHANGES. This was based on diffs to 7.3-devel and may not be applicable to 7.2. I have included a change covered by a previous bugfix patch I submitted (the problem with -.1 not being accepted by cube_in). It does not include a fix for the potential buffer overrun issue I reported for cube_yyerror in cubeparse.y.
Bruno Wolff III

+++ contrib/cube/cubedata.h
@@ -2,5 +2,5 @@ typedef struct NDBOX
 {
 	unsigned int size;			/* required to be a Postgres varlena type */
 	unsigned int dim;
-	float		x[1];
+	double		x[1];
 }	NDBOX;
